test(auth): organize server function fixtures
