+++ test/auth/server-functions/logout/next.tsx
@@ -1,9 +1,9 @@
 'use client'
 
-import { LogoutButton } from './logout.js'
+import { LogoutButton } from './index.js'
 
 const logoutFunction = async () => {
-  const { logoutFunction } = await import('./logoutFunction.js')
+  const { logoutFunction } = await import('./nextFunction.js')
 
   return logoutFunction()
 }
