+++ test/auth/server-functions/logout/nextFunction.tsx
@@ -2,7 +2,7 @@
 
 import { logout } from '@payloadcms/next/auth'
 
-import config from '../config.js'
+import config from '../../config.js'
 
 export async function logoutFunction() {
   try {
